Tidy up and refactor
